+++ src/function.js
@@ -11,3 +11,4 @@ export * from './function/until';
 export * from './function/fnOrError';
 export * from './function/noop';
 export * from './function/trampoline';
+export * from './function/toFunction';